requirements.txt
pytube
django-crispy-forms
crispy-bootstrap5
environs[django]

# Celery
# django-celery-beat 
# django-celery-results
# redis
